organized components to more logical directories
