+++ PennyPincherUI/src/app/app.component.html
@@ -1,21 +1,5 @@
 <main class="main">
-  <a
-    routerLink="/dashboard"
-    routerLinkActive="active"
-    ariaCurrentWhenActive="page"
-  >
-    <div>
-      <header class="brand-name">
-        <img
-          class="brand-logo"
-          src="/assets/homeIcon.svg"
-          alt="logo"
-          aria-hidden="true"
-        />
-      </header>
-      <h1>Penny Pincher 🪙</h1>
-    </div>
-  </a>
-
+  <app-header></app-header>
   <router-outlet />
+  <app-nav-bar></app-nav-bar>
 </main>
