+++ PennyPincherUI/src/app/app.component.ts
@@ -1,14 +1,11 @@
 import { Component, inject } from "@angular/core";
-import {
-	RouterModule,
-	RouterOutlet,
-	RouterLink,
-	RouterLinkActive,
-} from "@angular/router";
+import { RouterModule, RouterOutlet } from "@angular/router";
+import { NavBarComponent } from "./shared/nav-bar/nav-bar.component";
+import { HeaderComponent } from "./shared/header/header.component";
 
 @Component({
 	selector: "app-root",
-	imports: [RouterModule, RouterOutlet, RouterLink, RouterLinkActive],
+	imports: [RouterModule, RouterOutlet, HeaderComponent, NavBarComponent],
 	templateUrl: "./app.component.html",
 	styleUrl: "./app.component.css",
 })
