+++ PennyPincherUI/src/app/cash-flows/cash-flows.component.html
@@ -1,2 +1,1 @@
 <p>cash-flows works!</p>
-<app-nav-bar></app-nav-bar>
