+++ PennyPincherUI/src/app/dashboard/dashboard.component.html
@@ -1,2 +1,1 @@
 <p>dashboard works!</p>
-<app-nav-bar></app-nav-bar>
