+++ PennyPincherUI/src/app/expenses/expense.component.html
@@ -8,4 +8,3 @@
     <expense-form />
   </div>
 </div>
-<app-nav-bar></app-nav-bar>
