+++ PennyPincherUI/src/app/expenses/expense.component.ts
@@ -1,11 +1,10 @@
 import { Component, inject } from "@angular/core";
 import { ExpenseTable } from "./expense-table.component";
 import { ExpenseForm } from "./expense-form.component";
-import { NavBarComponent } from "../nav-bar/nav-bar.component";
 
 @Component({
 	selector: "app-expense",
-	imports: [ExpenseTable, ExpenseForm, NavBarComponent],
+	imports: [ExpenseTable, ExpenseForm],
 	templateUrl: "./expense.component.html",
 	styleUrl: "./expense.component.css",
 })
