+++ PennyPincherUI/src/app/shared/header/header.component.html
@@ -1,0 +1,17 @@
+<a
+  routerLink="/dashboard"
+  routerLinkActive="active"
+  ariaCurrentWhenActive="page"
+>
+  <div>
+    <header class="brand-name">
+      <img
+        class="brand-logo"
+        src="/assets/homeIcon.svg"
+        alt="logo"
+        aria-hidden="true"
+      />
+    </header>
+    <h1>Penny Pincher 🪙</h1>
+  </div>
+</a>
