+++ PennyPincherUI/src/app/shared/header/header.component.ts
@@ -1,0 +1,11 @@
+import { Component } from "@angular/core";
+import { RouterLink, RouterLinkActive } from "@angular/router";
+import { NavBarComponent } from "../nav-bar/nav-bar.component";
+
+@Component({
+	selector: "app-header",
+	imports: [RouterLink, RouterLinkActive, RouterLinkActive, NavBarComponent],
+	templateUrl: "./header.component.html",
+	styleUrl: "./header.component.css",
+})
+export class HeaderComponent {}
