+++ PennyPincherUI/src/app/shared/nav-bar/nav-bar.component.spec.ts
@@ -1,0 +1,22 @@
+import { type ComponentFixture, TestBed } from "@angular/core/testing";
+
+import { NavBarComponent } from "./nav-bar.component";
+
+describe("NavBarComponent", () => {
+	let component: NavBarComponent;
+	let fixture: ComponentFixture<NavBarComponent>;
+
+	beforeEach(async () => {
+		await TestBed.configureTestingModule({
+			imports: [NavBarComponent],
+		}).compileComponents();
+
+		fixture = TestBed.createComponent(NavBarComponent);
+		component = fixture.componentInstance;
+		fixture.detectChanges();
+	});
+
+	it("should create", () => {
+		expect(component).toBeTruthy();
+	});
+});
