+++ PennyPincherUI/src/app/shared/nav-bar/nav-bar.component.ts
